Get correct meeting ID from db and pass it to PDF generator

+++ app/src/main/java/com/vanigent/meetingapp/data/local/entity/MeetingEntity.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "meeting", indices = [Index(value = ["id"], unique = true)])
 data class MeetingEntity(
 
-    @PrimaryKey val id: Int? = null,
+    @PrimaryKey val id: Long? = null,
     @ColumnInfo(name = "location")
     val address: AddressEntity,
     @ColumnInfo(name = "coordinator_will_consume_food")
